+++ src/main/java/insane96mcp/enhancedai/modules/base/feature/Targeting.java
@@ -10,6 +10,7 @@ import insane96mcp.insanelib.util.IdTagMatcher;
 import insane96mcp.insanelib.util.MCUtils;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.Mob;
+import net.minecraft.world.entity.ai.attributes.AttributeModifier;
 import net.minecraft.world.entity.ai.attributes.Attributes;
 import net.minecraft.world.entity.ai.goal.Goal;
 import net.minecraft.world.entity.ai.goal.WrappedGoal;
@@ -18,18 +19,21 @@ import net.minecraft.world.entity.monster.Endermite;
 import net.minecraft.world.entity.monster.Spider;
 import net.minecraft.world.entity.player.Player;
 import net.minecraftforge.common.ForgeConfigSpec;
+import net.minecraftforge.common.ForgeMod;
 import net.minecraftforge.event.entity.EntityJoinWorldEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
+import java.util.UUID;
 import java.util.function.Predicate;
 
 @Label(name = "Targeting", description = "Change how mobs target players")
 public class Targeting extends Feature {
 
 	private final ForgeConfigSpec.ConfigValue<Integer> followRangeConfig;
+	private final ForgeConfigSpec.ConfigValue<Double> swimSpeedMultiplierConfig;
 	private final ForgeConfigSpec.ConfigValue<Double> xrayConfig;
 	private final ForgeConfigSpec.ConfigValue<Boolean> instaTargetConfig;
 	private final BlacklistConfig entityBlacklistConfig;
@@ -37,6 +41,7 @@ public class Targeting extends Feature {
 	private final List<String> entityBlacklistDefault = Arrays.asList("minecraft:enderman");
 
 	public int followRange = 64;
+	public double swimSpeedMultiplier = 2.5d;
 	public double xray = 0.15d;
 	public boolean instaTarget = true;
 	public ArrayList<IdTagMatcher> entityBlacklist;
@@ -46,8 +51,11 @@ public class Targeting extends Feature {
 		super(Config.builder, module);
 		Config.builder.comment(this.getDescription()).push(this.getName());
 		followRangeConfig = Config.builder
-				.comment("How far away can the mobs see the player. This overrides the vanilla value (16 for most mobs). Setting to 0 will leave the follow range as vanilla.")
+				.comment("How far away can the mobs see the player. This overrides the vanilla value (16 for most mobs). Setting to 0 will leave the follow range as vanilla. I recommend using mods like Mobs Properties Randomness to have more control over the attribute.")
 				.defineInRange("Follow Range Override", this.followRange, 0, 128);
+		swimSpeedMultiplierConfig = Config.builder
+				.comment("How faster mobs can swim. I recommend using mods like Mobs Properties Randomness to have more control over the attribute.")
+				.defineInRange("Swim Speed Multiplier", this.swimSpeedMultiplier, 0d, 4d);
 		xrayConfig = Config.builder
 				.comment("Chance for a mob to be able to see players through blocks.")
 				.defineInRange("XRay Chance", xray, 0d, 1d);
@@ -62,11 +70,14 @@ public class Targeting extends Feature {
 	public void loadConfig() {
 		super.loadConfig();
 		this.followRange = this.followRangeConfig.get();
+		this.swimSpeedMultiplier = this.swimSpeedMultiplierConfig.get();
 		this.xray = this.xrayConfig.get();
 		this.instaTarget = this.instaTargetConfig.get();
 		this.entityBlacklist = (ArrayList<IdTagMatcher>) IdTagMatcher.parseStringList(this.entityBlacklistConfig.listConfig.get());
 		this.entityBlacklistAsWhitelist = this.entityBlacklistConfig.listAsWhitelistConfig.get();
 	}
+
+	final UUID UUID_SWIM_SPEED_MULTIPLIER = UUID.fromString("6d2cb27e-e5e3-41b9-8108-f74131a90cce");
 
 	@SubscribeEvent
 	public void onMobSpawn(EntityJoinWorldEvent event) {
@@ -94,6 +105,10 @@ public class Targeting extends Feature {
 
 		if (this.followRange != 0) {
 			MCUtils.setAttributeValue(mobEntity, Attributes.FOLLOW_RANGE, this.followRange);
+		}
+
+		if (this.swimSpeedMultiplier != 0d) {
+			MCUtils.applyModifier(mobEntity, ForgeMod.SWIM_SPEED.get(), UUID_SWIM_SPEED_MULTIPLIER, "Enhanced AI Swim Speed Multiplier", this.swimSpeedMultiplier, AttributeModifier.Operation.MULTIPLY_BASE, false);
 		}
 
 		boolean hasTargetGoal = false;